[AI-57] Fixed flujo PRO

+++ app/controllers/registro_proveedores_controller.rb
@@ -175,24 +175,30 @@ class RegistroProveedoresController < ApplicationController
   def update
     @editar = params[:editar]
     @registro_proveedor = RegistroProveedor.find(params[:id])
+    puts "--------> #{@registro_proveedor.inspect}"
     asociar_institucion = @registro_proveedor.asociar_institucion
     contribuyente_id = @registro_proveedor.contribuyente_id
     rut_institucion = @registro_proveedor.rut_institucion
     registro_proveedores_params_sin_editar = registro_proveedores_params.except(:editar)
-
     respond_to do |format|
       if @registro_proveedor.update(registro_proveedores_params_sin_editar)
-        RegistroProveedor::UpdateService.new(@registro_proveedor, registro_proveedores_params, asociar_institucion, contribuyente_id, rut_institucion).perform
-
-        if params[:registro_proveedor][:editar] != 'true'
-          flujo = Flujo.where(id: 1000, contribuyente_id: 1000, tipo_instrumento_id: 26).first_or_create
-          tarea_pendiente = TareaPendiente.where(flujo_id: flujo.id, user_id: current_user.id).first
-          tarea_pendiente.destroy
-          tarea = Tarea.where(nombre: "PRO-003").first
-          TareaPendiente.create(flujo_id: flujo.id, tarea_id: tarea.id, estado_tarea_pendiente_id: EstadoTareaPendiente::NO_INICIADA, user_id: @registro_proveedor.user_encargado, data: @registro_proveedor.id)
-        end
+        if registro_proveedores_params_sin_editar[:carta_compromiso].nil?
+          RegistroProveedor::UpdateService.new(@registro_proveedor, registro_proveedores_params, asociar_institucion, contribuyente_id, rut_institucion).perform
+       
+          if params[:registro_proveedor][:editar] != 'true'
+            flujo = Flujo.where(id: 1000, contribuyente_id: 1000, tipo_instrumento_id: 26).first_or_create
+            tarea_pendiente = TareaPendiente.where(flujo_id: flujo.id, user_id: current_user.id).first
+            tarea_pendiente.destroy
+            tarea = Tarea.where(nombre: "PRO-003").first
+            TareaPendiente.create(flujo_id: flujo.id, tarea_id: tarea.id, estado_tarea_pendiente_id: EstadoTareaPendiente::NO_INICIADA, user_id: @registro_proveedor.user_encargado, data: @registro_proveedor.id)
+          end
+       end
         format.js {
           render js: "window.location='#{root_path}'"
+          flash[:success] = "Registro enviado correctamente"
+        }
+        format.html {
+          redirect_to root_path
           flash[:success] = "Registro enviado correctamente"
         }
        
@@ -359,10 +365,13 @@ class RegistroProveedoresController < ApplicationController
     respond_to do |format|
       if @registro_proveedor.update(registro_proveedores_params)
         @registro_proveedor.update(estado: 9)
-        # flujo = Flujo.where(registro_proveedor_id: @registro_proveedor.id).first
-        # tarea_pendiente = TareaPendiente.where(flujo_id: flujo.id).first
-        # tarea = Tarea.where(nombre: "PRO-008").first
-        # tarea_pendiente.update(tarea_id: tarea.id, user_id: @registro_proveedor.user_encargado)
+        flujo = Flujo.where(id: 1000, contribuyente_id: 1000, tipo_instrumento_id: 26).first_or_create
+        tarea = Tarea.where(nombre: "PRO-008").first
+        user = User.where(rut: @registro_proveedor.rut).first
+        TareaPendiente.create(flujo_id: flujo.id, tarea_id: tarea.id, estado_tarea_pendiente_id: EstadoTareaPendiente::NO_INICIADA, user_id: @registro_proveedor.user_encargado, data: @registro_proveedor.id)
+        tarea_previa = Tarea.where(nombre: 'PRO-007').first
+        tarea_pendiente = TareaPendiente.where(flujo_id: flujo.id, tarea_id: tarea_previa.id, estado_tarea_pendiente_id: EstadoTareaPendiente::NO_INICIADA, data: @registro_proveedor.id)
+        tarea_pendiente.first.delete
         if @registro_proveedor.region.present? && @registro_proveedor.comuna.present?
           @registro_proveedor.region = Region.find(@registro_proveedor.region.to_i).nombre
           @registro_proveedor.comuna = Comuna.find(@registro_proveedor.comuna.to_i).nombre
@@ -401,13 +410,20 @@ class RegistroProveedoresController < ApplicationController
 
       if value == 1
         @registro_proveedor.update!(estado: 9)
+        flujo = Flujo.where(id: 1000, contribuyente_id: 1000, tipo_instrumento_id: 26).first_or_create
+        tarea = Tarea.where(nombre: "PRO-009").first
+        user = User.where(rut: @registro_proveedor.rut).first
+        TareaPendiente.create(flujo_id: flujo.id, tarea_id: tarea.id, estado_tarea_pendiente_id: EstadoTareaPendiente::NO_INICIADA, user_id: @registro_proveedor.user_encargado, data: @registro_proveedor.id)
+        tarea_previa = Tarea.where(nombre: 'PRO-008').first
+        tarea_pendiente = TareaPendiente.where(flujo_id: flujo.id, tarea_id: tarea_previa.id, estado_tarea_pendiente_id: EstadoTareaPendiente::NO_INICIADA, data: @registro_proveedor.id)
+        tarea_pendiente.first.delete
       end
 
       if value == 2
         @registro_proveedor.update!(estado: 8)
         RegistroProveedorMailer.revision_negativa(@registro_proveedor).deliver_later
-        flujo = Flujo.where(registro_proveedor_id: @registro_proveedor.id).first
-        tarea_pendiente = TareaPendiente.where(flujo_id: flujo.id).first
+        flujo = Flujo.where(id: 1000, contribuyente_id: 1000, tipo_instrumento_id: 26).first_or_create
+        tarea_pendiente = TareaPendiente.where(flujo_id: 1000).first
         tarea_pendiente.destroy
         # flujo.destroy
       end
@@ -422,9 +438,10 @@ class RegistroProveedoresController < ApplicationController
       @registro_proveedor = RegistroProveedor.find(key)
 
       if @registro_proveedor.estado == 'actualizado'
-        @registro_proveedor.update!(fecha_revalidacion: value)
+        @registro_proveedor.update!(fecha_revalidacion: value, estado: 'aprobado')
         RegistroProveedorMailer.revision_positiva(@registro_proveedor).deliver_now
       end
+    @registro_proveedor.update!(fecha_revalidacion: value, estado: 'aprobado')
 
     end
 
@@ -483,6 +500,20 @@ class RegistroProveedoresController < ApplicationController
 
     redirect_to root_path
     flash[:success] = "Registro enviado correctamente"
+
+  end
+
+  #PRO-010
+  def enviar_carta_compromiso
+    @registro_proveedor = RegistroProveedor.find(params[:id])
+    @descargables_tarea = DescargableTarea.where(tarea_id: 104)
+
+  end
+
+  #PRO-010
+  def enviar_carta_compromiso
+    @registro_proveedor = RegistroProveedor.find(params[:id])
+    @descargables_tarea = DescargableTarea.where(tarea_id: 104)
 
   end
 
@@ -572,7 +603,7 @@ class RegistroProveedoresController < ApplicationController
   def registro_proveedores_params
     params.require(:registro_proveedor).permit(:rut, :nombre, :apellido, :email, :telefono, :profesion, :direccion, :region, :comuna, :ciudad, :asociar_institucion, :tipo_contribuyente_id, :terminos_y_servicion,
       :rut_institucion, :nombre_institucion, :tipo_contribuyente, :editar, :tipo_proveedor_id, :direccion_casa_matriz, :region_casa_matriz, :comuna_casa_matriz, :ciudad_casa_matriz, :contribuyente_id, :respuesta_comentario,
-      :archivo_respuesta_rechazo, :comentario_directiva, :respuesta_comentario_directiva, :archivo_respuesta_rechazo_directiva, :fecha_aprobado, :fecha_actualizado, :archivo_aprobado_directiva, :archivo_aprobado_directiva_cache,
+      :archivo_respuesta_rechazo, :comentario_directiva, :respuesta_comentario_directiva, :archivo_respuesta_rechazo_directiva, :fecha_aprobado, :fecha_actualizado, :archivo_aprobado_directiva, :archivo_aprobado_directiva_cache, :carta_compromiso, :carta_compromiso_cache,
       certificado_proveedores_attributes: [:id, :materia_sustancia_id, :actividad_economica_id,  :archivo_certificado, :archivo_certificado_cache, :_destroy], documento_registro_proveedores_attributes: [:id, :description, :archivo, :archivo_cache, :_destroy],
       certificado_proveedor_extras_attributes: [:id, :materia_sustancia_id, :actividad_economica_id, :archivo, :archivo_cache, :_destroy], documento_proveedor_extras_attributes: [:id, :description, :archivo, :archivo_cache, :_destroy])
   end
